wc minimal *almost* works [ci skip]
