+++ examples/wordcount/bin/wordcount_minimal.py
@@ -37,7 +37,9 @@ class Mapper(api.Mapper):
 class Reducer(api.Reducer):
 
     def reduce(self, context):
-        context.emit(context.key, sum(context.values))
+        # context.emit(context.key, sum(context.values))
+        v = sum(context.values)
+        context.emit(context.key.encode("utf-8"), str(v).encode("utf-8"))
 
 
 FACTORY = pipes.Factory(mapper_class=Mapper, reducer_class=Reducer)
